Organise performance check using `*.performance.js`
Add `getTimes` to `time.js` to get `*.performance.js` file and compare functions
