+++ TIMES.md
@@ -1,7 +1,7 @@
 ## Nanoutils method's time
 | Method | Lib   |       0% |      50% |    100% |
 | ------ | ----- | -------- | -------- | ------- |
-| union  | nano  |  16.39ms |  29.15ms |  8.95ms |
-|        | ramda |   1.49ms |   0.52ms |  0.56ms |
-|        | diff  | +14.89ms | +28.63ms | +8.38ms |
+| union  | nano  |  21.26ms |  37.87ms |  9.80ms |
+|        | ramda |   1.02ms |   0.68ms |  0.60ms |
+|        | diff  | +20.24ms | +37.19ms | +9.20ms |
 | ------ | ----- | -------- | -------- | ------- |
